backend/requirements.txt
# backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-ckeditor==6.5.1
django-redis==5.3.0
googletrans==3.1.0a0
celery==5.2.7
pytest==7.4.3
flake8==6.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.0
pytest-django